Added compact fuel prices display
The change adds a boolean setting "Compact fuel display", disabled by default. Once enabled, it changes the display of the fuel prices in the Docks and hides the list of prices behind the HTML details/summary element. This helps to not clutter the page with extra data when it's not needed and especially helps folks who have career tasks in the Docks.