fuel-price-strategy: further disambiguation
Changes integrated from traktofon/tau-utility-scripts.

For stations with converged fuel price, we can calculate the current
item price, and if that doesn't match the min or max from the current
price range, this reduces the number of candidate stations for the
min/max -- potentially to just one.  Double-check that the fuel price
thus determined is inside the interval found in the first phase.

TODO: tolerance in `equals_approx` might still need to be raised

Other changes:
- replace use of `urllib.request` by `requests`, as that's
  already used anyway